+++ site Scrum do Zero/public/assets/css/style.css
@@ -1,2 +1,3 @@
+@import url(home.css);
+@import url(quizPage.css);
 @import url(global.css);
-@import url(home.css)
