+++ site Scrum do Zero/src/server.js
@@ -22,10 +22,14 @@ app.use('/assets', express.static(path.join(publicPath,"assets")))
 
 app.use("/api", router)
 
-app.use(function(_req,res){
-    res.redirect("not-found.html");
-});
-
 app.listen(PORT, function () {
     console.log(`Rodando em http://localhost:${PORT}`)
 })
+
+app.get("/quiz-page", function(_req, res){
+    res.sendFile(path.join(pagesPath, "quiz-page.html"));
+});
+
+app.use(function(_req,res){
+    res.redirect("not-found.html");
+});
