--- siteScrumdoZero/.env
@@ -1,10 +1,0 @@
-PORT=3000
-
-POSTGRES_HOST=localhost
-POSTGRES_USER=postgres
-POSTGRES_PASSWORD=123
-POSTGRES_DB=abp
-POSTGRES_PORT=5432
-
-JWT_SECRET=@123_56@
-DEFAULT_EXPIRES_IN_SECONDS=3600
