--- siteScrumdoZero/public/assets/css/home.css
@@ -1,409 +1,0 @@
-.page-main-section {
-    display: flex;
-    background: linear-gradient(135deg, #1a1a1a 0%, #2d2d2d 100%);
-    position: relative;
-    min-height: 100vh;
-}
-
-/* Desktop container - oculto no mobile */
-.main-container {
-    display: none;
-}
-
-.footer {
-    display: none;
-}
-
-/* Mobile View - Padrão */
-.mobile-view {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.mobile-article-section {
-    display: flex;
-    flex-direction: column;
-    align-items: stretch;
-    text-align: center;
-    max-width: clamp(300px, 80%, 450px);
-    min-height: 100vh;
-    padding: 2rem 0;
-}
-
-.mobile-logo-title {
-    font-size: 2rem;
-    font-weight: 700;
-    margin-bottom: 40px;
-}
-
-.mobile-logo {
-    width: 140px;
-    height: 140px;
-    margin: 0 auto 40px;
-    background: linear-gradient(135deg, #4a4a4a 0%, #6a6a6a 100%);
-    border-radius: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    position: relative;
-    box-shadow: 0 4px 15px rgba(0, 0, 0, 0.4);
-}
-
-.mobile-logo-inner {
-    width: 80%;
-    height: 80%;
-    border-radius: 50%;
-    background: #2a2a2a;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 42px;
-    font-weight: 700;
-}
-
-.mobile-logo-text {
-    position: absolute;
-    bottom: 18px;
-    font-size: 12px;
-    font-weight: 600;
-    letter-spacing: 2px;
-}
-
-.info-section {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 3rem;
-    margin-bottom: 4rem;
-}
-
-.info-section h2 {
-    font-size: 1.25rem;
-    font-weight: 600;
-}
-
-.info-section p {
-    font-size: 1rem;
-    line-height: 1.5;
-}
-
-.mobile-buttons {
-    display: flex;
-    gap: 1.5rem;
-    margin-top: auto;
-    justify-content: center;
-    flex-direction: column;
-}
-
-.mobile-button {
-    padding: 14px 35px;
-    background: transparent;
-    border: 2px solid white;
-    border-radius: 25px;
-    font-size: 16px;
-    font-weight: 600;
-    color: white;
-    cursor: pointer;
-    transition: all 0.3s;
-}
-
-.mobile-button:hover {
-    background: white;
-    color: #1a1a1a;
-}
-
-/* Estilos desktop - inicialmente ocultos */
-.left-section,
-.right-section {
-    display: none;
-}
-
-.mobile-login-section {
-    min-height: 100vh;
-    min-width: clamp(300px, 80%, 450px);
-    padding: 2rem 0;
-    align-items: center;
-    display: flex;
-    justify-content: center;
-}
-
-.login-card {
-    background: white;
-    /* border: 1px solid #e0e0e0; */
-    border-radius: 12px;
-    padding: 2rem clamp(14px, 6vw, 1.9rem);
-    width: 100%;
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.05);
-    color: black;
-}
-
-.login-header {
-    text-align: center;
-    margin-bottom: 25px;
-}
-
-.login-header h3 {
-    font-size: clamp(1.5rem, 2vh, 2rem);
-    font-weight: 600;
-    margin-bottom: 8px;
-    color: #333;
-}
-
-.login-header p {
-    font-size: 1rem;
-    color: #666;
-}
-
-.form-login{
-    display: flex;
-    flex-direction: column;
-    gap: 1.25rem;
-}
-
-.form-group {
-    display: flex;
-    flex-direction: column;
-    gap: 0.875rem;
-}
-
-.form-label {
-    display: block;
-    font-size: 1rem;
-    color: #333;
-    font-weight: 500;
-}
-
-.form-input {
-    width: 100%;
-    padding: 11px 14px;
-    border: 1px solid #d0d0d0;
-    border-radius: 6px;
-    font-size: 13px;
-    transition: border-color 0.3s;
-}
-
-.form-input:focus {
-    outline: none;
-    border-color: #333;
-}
-
-.form-input::placeholder {
-    color: #999;
-    font-size: 12px;
-}
-
-.login-button {
-    width: 100%;
-    padding: 11px;
-    background: white;
-    border: 2px solid #333;
-    border-radius: 25px;
-    font-size: 0.875rem;
-    font-weight: 600;
-    color: #333;
-    cursor: pointer;
-    transition: all 0.3s;
-}
-
-.login-button:hover {
-    background: #333;
-    color: white;
-}
-
-.register-link {
-    text-align: center;
-    margin-top: 18px;
-    font-size: 0.8rem;
-    color: #666;
-}
-
-.register-link a {
-    color: #333;
-    text-decoration: none;
-    font-weight: 600;
-}
-
-.register-link a:hover {
-    text-decoration: underline;
-}
-
-
-@media (min-width: 1024px) {
-
-    .page-main-section {
-        background: transparent;
-    }
-
-    .page-main-section::after {
-        display: none;
-    }
-
-    .mobile-view {
-        display: none;
-    }
-
-    .main-container {
-        display: flex;
-        flex: 1;
-    }
-
-    .left-section {
-        display: flex;
-        background: linear-gradient(135deg, #1a1a1a 0%, #2d2d2d 100%);
-        color: white;
-        flex: 1;
-        flex-direction: column;
-        position: relative;
-        overflow: hidden;
-        padding: 0 50px;
-        justify-content: center;
-    }
-
-    .left-section::after {
-        content: '';
-        position: absolute;
-        bottom: -100px;
-        left: -100px;
-        width: 400px;
-        height: 400px;
-        background: radial-gradient(circle, rgba(255, 255, 255, 0.03) 0%, transparent 70%);
-        border-radius: 50%;
-    }
-
-    .logo-container {
-        margin-bottom: 25px;
-        position: relative;
-        z-index: 1;
-    }
-
-    .logo-title {
-        font-size: 36px;
-        font-weight: 700;
-        margin-bottom: 20px;
-    }
-
-    .logo {
-        width: 100px;
-        height: 100px;
-        margin-bottom: 25px;
-        background: linear-gradient(135deg, #4a4a4a 0%, #6a6a6a 100%);
-        border-radius: 50%;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        position: relative;
-        box-shadow: 0 4px 15px rgba(0, 0, 0, 0.3);
-    }
-
-    .logo-inner {
-        width: 80%;
-        height: 80%;
-        border-radius: 50%;
-        background: #2a2a2a;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        font-size: 32px;
-        font-weight: 700;
-    }
-
-    .logo-text {
-        position: absolute;
-        bottom: 12px;
-        font-size: 10px;
-        font-weight: 600;
-        letter-spacing: 2px;
-    }
-
-    .content-wrapper {
-        position: relative;
-        z-index: 1;
-        max-width: 380px;
-    }
-
-    .main-title {
-        font-size: 24px;
-        font-weight: 600;
-        margin-bottom: 25px;
-        line-height: 1.4;
-    }
-
-    .description {
-        font-size: 16px;
-        margin-bottom: 20px;
-        line-height: 1.6;
-        color: #d4d4d4;
-    }
-
-    .note {
-        font-size: 14px;
-        margin-bottom: 25px;
-        line-height: 1.5;
-        color: #b4b4b4;
-    }
-
-    .link-section {
-        font-size: 14px;
-        color: #d4d4d4;
-    }
-
-    .link-section a {
-        color: #ffffff;
-        text-decoration: underline;
-    }
-
-    .right-section {
-        display: flex;
-        background: #f5f5f5;
-        flex: 1;
-        align-items: center;
-        justify-content: center;
-        padding: 40px;
-    }
-
-    .login-card {
-        padding: 40px 35px;
-        max-width: 400px;
-    }
-
-    .login-header {
-        margin-bottom: 30px;
-    }
-
-    .login-title {
-        font-size: 20px;
-    }
-
-    .login-subtitle {
-        font-size: 13px;
-    }
-
-    .form-group {
-        margin-bottom: 20px;
-    }
-
-    .form-label {
-        font-size: 13px;
-    }
-
-    .form-input {
-        padding: 12px 15px;
-        font-size: 14px;
-    }
-
-    .form-input::placeholder {
-        font-size: 13px;
-    }
-
-    .login-button {
-        padding: 12px;
-        font-size: 15px;
-    }
-
-    .register-link {
-        margin-top: 20px;
-        font-size: 13px;
-    }
-
-}
