--- siteScrumdoZero/public/assets/css/style.css
@@ -1,3 +1,0 @@
-@import url(home.css);
-@import url(quizPage.css);
-@import url(global.css);
