--- siteScrumdoZero/public/pages/index.html
@@ -1,138 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-BR">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="/assets/css/style.css">
-  <script defer src="/assets/js/home.js"></script>
-  <title>ABP</title>
-</head>
-
-<body>
-
-  <main class="page-main-section">
-
-    <div class="main-container">
-      <div class="left-section">
-        <div class="logo-container">
-          <h1 class="logo-title">Scrum do Zero</h1>
-          <div class="logo">
-            <div class="logo-inner">◐</div>
-            <div class="logo-text">SCRUM</div>
-          </div>
-        </div>
-
-        <div class="content-wrapper">
-          <h2 class="main-title">
-            Seu caminho para dominar<br />
-            metodologias ágeis começa aqui.
-          </h2>
-
-          <p class="description">
-            Plataforma de avaliações progressivas sobre Scrum, com questões Scrum, questões diversificadas,
-            acompanhamento do seu progresso e certificado ao final da jornada.
-          </p>
-
-          <p class="note">
-            Desejo relembrar e/ou estudar um pouco mais antes de realizar o desafio?
-          </p>
-
-          <div class="link-section">
-            Acesse o link: <a href="https://linq.com/klpsv2v" target="_blank"
-              rel="noopener noreferrer">https://linq.com/klpsv2v</a>
-          </div>
-        </div>
-      </div>
-
-      <div class="right-section">
-        <div class="login-card">
-          <div class="login-header">
-            <h3 class="login-title">Bem vindo!</h3>
-            <p class="login-subtitle">Faça login para continuar sua jornada</p>
-          </div>
-
-          <form>
-            <div class="form-group">
-              <label class="form-label">Seu CPF</label>
-              <input type="text" class="form-input" />
-            </div>
-
-            <div class="form-group">
-              <label class="form-label">Sua senha</label>
-              <input type="password" class="form-input" placeholder="Esqueci minha senha" />
-            </div>
-
-            <button type="submit" class="login-button">Login</button>
-          </form>
-
-          <div class="register-link">
-            Não tem uma conta? <a href="#cadastrar">Cadastrar-se</a>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <section class="mobile-view">
-
-      <article class="mobile-article-section">
-        <h1 class="mobile-logo-title">Scrum do Zero</h1>
-
-        <div class="mobile-logo">
-          <div class="mobile-logo-inner">◐</div>
-          <div class="mobile-logo-text">SCRUM</div>
-        </div>
-
-        <div class="info-section">
-          <h2>
-            Seu caminho para dominar metodologias ágeis começa aqui.
-          </h2>
-
-          <p>
-            Plataforma de avaliações progressivas sobre Scrum, com questões Scrum, questões diversificadas,
-            acompanhamento
-            do seu progresso e certificado ao final da jornada.
-          </p>
-        </div>
-
-        <div class="mobile-buttons">
-          <a href="#login-card-sec" class="mobile-button" id="access-btn">Acessar</a>
-          <button class="mobile-button">Cadastrar</button>
-        </div>
-      </article>
-
-      <article class="mobile-login-section hidden" id="login-card-sec">
-        <div class="login-card">
-          <div class="login-header">
-            <h3>Bem vindo!</h3>
-            <p>Faça login para continuar sua jornada</p>
-          </div>
-      
-          <form class="form-login">
-            <div class="form-group">
-              <label for="cpf-area" class="form-label">Seu CPF</label>
-              <input class="form-input" name="CPF" data-required="This field is required" id="cpf-area" type="text" />
-            </div>
-      
-            <div class="form-group">
-              <label for="password-area" class="form-label">Sua senha</label>
-              <input class="form-input" name="password" data-required="This field is required" id="password-area" type="password" >
-              <a href="">Esqueci minha senha</a>
-            </div>
-      
-            <button type="submit" class="login-button">Login</button>
-          </form>
-      
-          <div class="register-link">
-            Não tem uma conta? <a href="#cadastrar">Cadastrar-se</a>
-          </div>
-        </div>
-      </article>
-
-    </section>
-
-  </main>
-
-</body>
-
-</html>
