--- siteScrumdoZero/public/pages/not-found.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-BR">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>error</title>
-</head>
-<body>
-    <div>Pagina não encontrada</div>
-</body>
-</html>
