--- siteScrumdoZero/src/infra/init/01_schema_modulos.sql
@@ -1,4 +1,0 @@
-CREATE TABLE IF NOT EXISTS public.modulos (
-  id_modulo INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-  titulo VARCHAR(255) NOT NULL
-);
