--- siteScrumdoZero/src/infra/init/02_schema_questoes.sql
@@ -1,25 +1,0 @@
-CREATE TABLE IF NOT EXISTS public.questoes (
-  id_questao INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-  id_modulo INTEGER NOT NULL,
-  grupo INTEGER,
-  numero INTEGER,
-  dificuldade VARCHAR(20),
-  enunciado TEXT,
-  alternativa_correta CHAR(1),
-  alternativa_a TEXT,
-  alternativa_b TEXT,
-  alternativa_c TEXT,
-  alternativa_d TEXT,
-  imagem VARCHAR(255),
-  CONSTRAINT fk_questoes_modulo
-    FOREIGN KEY (id_modulo)
-    REFERENCES public.modulos (id_modulo)
-    ON UPDATE NO ACTION
-    ON DELETE NO ACTION
-);
-
-CREATE INDEX IF NOT EXISTS idx_questoes_modulo_grupo
-  ON public.questoes (id_modulo, grupo);
-
-CREATE INDEX IF NOT EXISTS idx_questoes_dificuldade
-  ON public.questoes (dificuldade);
