--- siteScrumdoZero/src/infra/init/03_schema_usuarios.sql
@@ -1,20 +1,0 @@
-CREATE TABLE IF NOT EXISTS public.usuarios (
-  id_usuario INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-  nome VARCHAR(100),
-  email VARCHAR(100),
-  cpf VARCHAR(11) NOT NULL,
-  senha VARCHAR(200),
-  certificado_hash VARCHAR(96)
-);
-
-CREATE UNIQUE INDEX IF NOT EXISTS uq_usuarios_cpf
-  ON public.usuarios (cpf)
-  WHERE cpf IS NOT NULL;
-
-CREATE UNIQUE INDEX IF NOT EXISTS uq_usuarios_email
-  ON public.usuarios (email)
-  WHERE email IS NOT NULL;
-
-CREATE UNIQUE INDEX IF NOT EXISTS uq_usuarios_certificado_hash
-  ON public.usuarios (certificado_hash)
-  WHERE certificado_hash IS NOT NULL;
