--- siteScrumdoZero/src/infra/init/04_schema_exames.sql
@@ -1,17 +1,0 @@
-CREATE TABLE IF NOT EXISTS public.exames (
-  id_exame INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-  id_modulo INTEGER NOT NULL,
-  id_usuario INTEGER NOT NULL,
-  grupo INTEGER,
-  tentativa INTEGER,
-  CONSTRAINT fk_exames_usuario
-    FOREIGN KEY (id_usuario)
-    REFERENCES public.usuarios (id_usuario)
-    ON DELETE NO ACTION
-    ON UPDATE NO ACTION,
-  CONSTRAINT fk_exames_modulo
-    FOREIGN KEY (id_modulo)
-    REFERENCES public.modulos (id_modulo)
-    ON DELETE NO ACTION
-    ON UPDATE NO ACTION
-);
