--- siteScrumdoZero/src/infra/init/05_schema_respostas.sql
@@ -1,24 +1,0 @@
-CREATE TABLE IF NOT EXISTS public.respostas (
-  id_resposta INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-  id_exame INTEGER NOT NULL,
-  id_questao INTEGER NOT NULL,
-  nota INTEGER,
-  resposta CHAR(1),
-  respondido_em TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP,
-  CONSTRAINT fk_respostas_questao
-    FOREIGN KEY (id_questao)
-    REFERENCES public.questoes (id_questao)
-    ON DELETE NO ACTION
-    ON UPDATE NO ACTION,
-  CONSTRAINT fk_respostas_exame
-    FOREIGN KEY (id_exame)
-    REFERENCES public.exames (id_exame)
-    ON DELETE NO ACTION
-    ON UPDATE NO ACTION
-);
-
-CREATE INDEX IF NOT EXISTS idx_respostas_exame
-  ON public.respostas (id_exame);
-
-CREATE INDEX IF NOT EXISTS idx_respostas_questao
-  ON public.respostas (id_questao);
