--- siteScrumdoZero/src/infra/init/seed-data/modulos.csv
@@ -1,6 +1,0 @@
-id_modulo;titulo
-1;Fundamentos das Metodologias Ágeis
-2;Scrum: Estrutura, Papéis e Artefatos
-3;Eventos do Scrum e Fluxo de Trabalho
-4;Práticas Ágeis, Métricas e Qualidade
-5;Aplicação Prática, Cenários e Análise Crítica
