--- siteScrumdoZero/src/routes/auth.routes.js
@@ -1,34 +1,0 @@
-const { Router } = require("express")
-const {findUsuarioByCpfAndSenha} = require("../repositories/usuarios.repositories")
-const { createToken } = require("../utils/jwt")
-
-const router = Router()
-
-/* POST LOGIN USUÁRIO
-curl -X POST http://localhost:3000/api/auth/login \
-  -H "Content-Type: application/json" \
-  -d '{"cpf":"12345678901", "senha":"123456"}'
-*/
-
-router.post("/login", async function (req, res) {
-    const { cpf, senha } = req.body
-         if (!cpf || !senha) {
-        return res.status(400)
-            .json({ message: "CPF ou senha inválidos" })
-    }
-    try{
-        const usuario = await findUsuarioByCpfAndSenha(cpf, senha)
-        const token = createToken({ id_usuario: usuario.id_usuario})
-        return res.status(200).json({
-            token,
-            nome: usuario.nome
-        })
-    }catch(e){
-        return res.status(500).json({
-                message: e.message
-            })
-    }
-})
-
-
-module.exports = router;
