--- siteScrumdoZero/src/routes/index.js
@@ -1,14 +1,0 @@
-const { Router } = require("express")
-const usuarios = require("./usuarios.routes")
-const auth = require("./auth.routes")
-
-const router = Router()
-
-router.use("/usuarios", usuarios)
-router.use("/auth", auth)
-
-router.use(function(_req,res){
-    res.status(404).json({message: "Rota inexistente"})
-})
-
-module.exports = router;
