--- siteScrumdoZero/src/server.js
@@ -1,35 +1,0 @@
-const express = require('express');
-const dotenv = require('dotenv');
-const path = require('path');
-const router = require("./routes") 
-
-dotenv.config({
-    quiet: true,
-    path: path.resolve(__dirname, "..", ".env")
-});
-
-const app = express();
-app.use(express.json());
-
-const PORT = process.env.PORT
-const publicPath = path.join(__dirname, "..", "public")
-const pagesPath = path.join(publicPath, "pages")
-const assetsPath = path.join(publicPath, "assets")
-
-app.use("/", express.static(pagesPath))
-app.use("/assets", express.static(assetsPath))
-app.use('/assets', express.static(path.join(publicPath,"assets")))
-
-app.use("/api", router)
-
-app.listen(PORT, function () {
-    console.log(`Rodando em http://localhost:${PORT}`)
-})
-
-app.get("/quiz-page", function(_req, res){
-    res.sendFile(path.join(pagesPath, "quiz-page.html"));
-});
-
-app.use(function(_req,res){
-    res.redirect("not-found.html");
-});
