--- siteScrumdoZero/src/utils/password.js
@@ -1,26 +1,0 @@
-const {randomBytes, scryptSync, timingSafeEqual} = require("crypto")
-
-function hashPassword(password){
-    const salt = randomBytes(16).toString("hex")
-    const hash = scryptSync(password, salt, 64).toString("hex")
-    return `${ salt }:${ hash }`
-}
-
-function verifyPassword(password, storedPassword){
-    const [salt, storedHash] = (storedPassword || "").split(":")
-    if (!salt || !storedHash) {
-        return false
-    }
-    const hash = scryptSync(password, salt, 64)
-    const storedHashBuffer = Buffer.from(storedHash, "hex")
-
-    if (hash.length !== storedHashBuffer.length) {
-        return false
-    }
-    return timingSafeEqual(hash, storedHashBuffer)
-}
-
-module.exports = {
-    hashPassword,
-    verifyPassword
-}
